% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.modelProperties}
\alias{.modelProperties}
\title{Fetch model properties for all models in a list}
\usage{
.modelProperties(models)
}
\arguments{
\item{models}{list of (lm) models for which to fetch properties}
}
\description{
Fetch model properties for all models in a list
}
