% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.modelValues}
\alias{.modelValues}
\title{Return a row with columns named for the coefficients and values containing
the requested property for each of models}
\usage{
.modelValues(models, property = "Estimate")
}
\arguments{
\item{models}{list of (lm) models to extract property from}

\item{property}{to extract}
}
\description{
Return a row with columns named for the coefficients and values containing
the requested property for each of models
}
