% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{kableANOVA}
\alias{kableANOVA}
\title{Format ANOVA results as a neat kableExtra printed table. Designed with
ezANOVA in mind.}
\usage{
kableANOVA(ANOVA, ...)
}
\arguments{
\item{ANOVA}{ANOVA results to format}

\item{...}{passed on to \code{kableExtra::kable}}
}
\value{
formatted version of ANOVA for inclusion in a report
}
\description{
Format ANOVA results as a neat kableExtra printed table. Designed with
ezANOVA in mind.
}
\examples{
\dontrun{
ezOutput <- ez::ezANOVA(...)
kableANOVA(ezOutput$ANOVA, caption = "Table of ANOVA results")
}

}
