% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{kableLM}
\alias{kableLM}
\title{Format LM results as a neat kableExtra printed table.}
\usage{
kableLM(LM, ...)
}
\arguments{
\item{LM}{lm results to format}

\item{...}{passed on to \code{kableExtra::kable}}
}
\value{
formatted version of LM for inclusion in a report
}
\description{
Format LM results as a neat kableExtra printed table.
}
\examples{
r <- lm(mpg ~ cyl * wt, data = mtcars)
kableLM(r)

}
