% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{kableModelComparison}
\alias{kableModelComparison}
\title{Produce a neat table showing comparisons between various models}
\usage{
kableModelComparison(models, ...)
}
\arguments{
\item{models}{list of lm models to compare with anova()}

\item{...}{passed on to \code{kableExtra::kable}}
}
\description{
Produce a neat table showing comparisons between various models
}
\examples{
m1 <- lm(Sepal.Length ~ Species, iris)
m2 <- lm(Sepal.Length ~ Species + Petal.Length, iris)
m3 <- lm(Sepal.Length ~ Species + Petal.Length + Petal.Width, iris)
m4 <- lm(Sepal.Length ~ Species + Petal.Length + Petal.Width + Sepal.Width, iris)
kableModelComparison(list(m1, m2, m3, m4))

}
