% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{md.BF}
\alias{md.BF}
\title{Extract the Bayes Factor from a Bayesian test}
\usage{
md.BF(bayesTest, ...)
}
\arguments{
\item{bayesTest}{the test whose parameter will be extracted}

\item{...}{Arguments passed on to \code{num2str}
\describe{
  \item{num}{number to convert to string}
  \item{precision}{decimal places to preserve}
  \item{isProportion}{whether to strip leading 0 for 0.x values}
  \item{truncateZeros}{whether to strip trailing 0s}
}}
}
\description{
Extract the Bayes Factor from a Bayesian test
}
\examples{
\dontrun{

data <- data.frame(x = rnorm(100), y = rnorm(100, 0.2)) # two normal distributions with some overlap
md.BF(BayesFactor::ttestBF(data$x, data$y, paired = T))
}

}
