% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{md.mean}
\alias{md.mean}
\title{Print the mean and CIs of a vector}
\usage{
md.mean(
  vector,
  label = "*M*",
  decimals = 2,
  na.rm = F,
  conf.int = 0.95,
  isProportion = F,
  showRange = F
)
}
\arguments{
\item{vector}{data in}

\item{label}{markdown prefix for the stats}

\item{decimals}{decimal places to round to}

\item{na.rm}{whether NA values are removed before averaging}

\item{conf.int}{width of the confidence intervals, NA to suppress}

\item{isProportion}{whether to print the values as proportions (strip leading 0)}

\item{showRange}{whether to include the range of the data}
}
\value{
string representation of the mean, CIs, and range of the \code{vector}
}
\description{
Print the mean and CIs of a vector
}
\examples{
md.mean(rnorm(1000), label = '*M*|random')
md.mean(rnorm(1000), label = '*M*|random', conf.int = NA) # no longer needs lsr package
md.mean(runif(1000), label = 'Mean probability', isProportion = TRUE, showRange = TRUE)

}
