% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{md.t}
\alias{md.t}
\title{Appropriately formatted t-test output}
\usage{
md.t(result, decimals = 2, decimals.p = NULL)
}
\arguments{
\item{result}{t.test output to format}

\item{decimals}{number of decimal places for means}

\item{decimals.p}{number of decimal places for p-values, defaults to
\code{decimals}+1}
}
\description{
Appropriately formatted t-test output
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100, 0.2)) # two normal distributions with some overlap
md.t(t.test(data$x, data$y, paired = TRUE), decimals = 3)

}
