% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{md.ttest}
\alias{md.ttest}
\title{Run a t-test and return neatly formatted output \code{x}, and either \code{y}
or \code{mu} must be specified.}
\usage{
md.ttest(x, y = NULL, labels = NULL, mu = NULL, paired = F, ...)
}
\arguments{
\item{x}{vector for testing}

\item{y}{additional vector for testing. If singular, functions as \code{mu}}

\item{labels}{vector of labels for x and y. If NULL, use attr(., 'label'), NA
to suppress this behaviour.}

\item{mu}{value against which x will be tested}

\item{paired}{whether observations are paired (e.g. repeated measures)}

\item{...}{
  Arguments passed on to \code{\link[=md.mean]{md.mean}}
  \describe{
    \item{\code{vector}}{data in}
    \item{\code{label}}{markdown prefix for the stats}
    \item{\code{conf.int}}{width of the confidence intervals, NA to suppress}
    \item{\code{na.rm}}{whether NA values are removed before averaging}
    \item{\code{decimals}}{decimal places to round to}
    \item{\code{isProportion}}{whether to print the values as proportions (strip leading 0)}
    \item{\code{showRange}}{whether to include the range of the data}
  }}
}
\value{
formatted string like t(df) = 3.68, p < .05, d = 0.23, BF = 4.50; M1
  = 3.05 [1.23, 4.55], M2 = 5.12 [4.20, 5.99]
}
\description{
Run a t-test and return neatly formatted output \code{x}, and either \code{y}
or \code{mu} must be specified.
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100, 0.2)) # two normal distributions with some overlap
tt <- md.ttest(data$x, data$y, c('*Mean|Control*', '*Mean|Treatment*'), paired = TRUE)
cat(tt)

}
