% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{num2str}
\alias{num2str}
\title{Format number to be a string beginning with a decimal point.}
\usage{
num2str(
  num,
  precision = 2,
  isProportion = F,
  truncateZeros = F,
  minPrefix = NA,
  ...
)
}
\arguments{
\item{num}{number to convert to string}

\item{precision}{decimal places to preserve}

\item{isProportion}{whether to strip leading 0 for 0.x values}

\item{truncateZeros}{whether to strip trailing 0s}

\item{minPrefix}{for transforming numbers like .000 into < .001}

\item{...}{discarded arguments used to allow overflowed calls from other functions}
}
\value{
\code{num} stripped of leading 0s and rounded to \code{precision} decimal places
}
\description{
Format number to be a string beginning with a decimal point.
}
\examples{
data.frame(input = c(seq(-10,10),100), output = num2str(exp(c(seq(-10,10),100)), 4))
data.frame(input = c(seq(-10,10),100), output = num2str(exp(c(seq(-10,10),100)), 4, isProportion = TRUE))

}
