% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{num2str.tibble}
\alias{num2str.tibble}
\title{Format entries in a tibble using num2str}
\usage{
num2str.tibble(tbl, precision = 2, isProportion = F,
  truncateZeros = F, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{num2str}
\describe{
  \item{num}{number to convert to string}
  \item{precision}{decimal places to preserve}
  \item{isProportion}{whether to strip leading 0 for 0.x values}
  \item{truncateZeros}{whether to strip trailing 0s}
}}
}
\description{
Parameters for \code{num2str} can be specified as vectors indicating the
columns to which values apply, with NA interpreted as using the default value
}
\examples{
library(tibble)
x <- tibble(chr = "character", int = 1:10, prop = runif(10), tProp = runif(10))
num2str.tibble(x, isProportion = c(NA, NA, T, T), truncateZeros = c(F, F, F, T))

}
