% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{p2str}
\alias{p2str}
\title{Wrapper for \code{num2str(..., isProportion = T, prefix = " = ", minPrefix = " < ")}
Designed for reporting p-values}
\usage{
p2str(x, precision = 3, prefix = " = ", minPrefix = " < ", ...)
}
\arguments{
\item{x}{number to convert to string}

\item{precision}{decimal places to preserve}

\item{prefix}{prefix for the number string (e.g. " = ")}

\item{minPrefix}{for transforming numbers like .000 into < .001}

\item{...}{
  Arguments passed on to \code{\link[=num2str]{num2str}}
  \describe{
    \item{\code{num}}{number to convert to string}
    \item{\code{isProportion}}{whether to strip leading 0 for 0.x values}
    \item{\code{truncateZeros}}{whether to strip trailing 0s}
  }}
}
\value{
\code{x} stripped of leading 0s and rounded to \code{precision} decimal places
}
\description{
Wrapper for \code{num2str(..., isProportion = T, prefix = " = ", minPrefix = " < ")}
Designed for reporting p-values
}
\examples{
data.frame(input = c(seq(-10,10),100),
  num = num2str(exp(c(seq(-10,10),100)), 4),
  p = p2str(exp(c(seq(-10,10),100)), 4)
)

}
