% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{lteq}
\alias{lteq}
\title{Format s with = if it's not 0, or < otherwise}
\usage{
lteq(s, non_equal_char = "<", sep = " ")
}
\arguments{
\item{s}{string to format}

\item{non_equal_char}{character to use if x != 0}

\item{sep}{separator between the relationship character and s}
}
\description{
Format s with = if it's not 0, or < otherwise
}
\examples{
lteq('.0001')
lteq('.000')

p <- t.test(rnorm(100, 1))$p.value
p
lteq(p)
# string format first!
lteq(prop2str(p, precision = 3))

}
