% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{md.ttestBF}
\alias{md.ttestBF}
\title{Return only the BayesFactor and means part of a t-test}
\usage{
md.ttestBF(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=md.ttest]{md.ttest}}
  \describe{
    \item{\code{x}}{vector for testing}
    \item{\code{y}}{additional vector for testing. If singular, functions as \code{mu}}
    \item{\code{labels}}{vector of labels for x and y. If NULL, use attr(., 'label'), NA
to suppress this behaviour.}
    \item{\code{mu}}{value against which x will be tested}
    \item{\code{paired}}{whether observations are paired (e.g. repeated measures)}
  }}
}
\description{
Return only the BayesFactor and means part of a t-test
}
